Moved winston configuration logic into server-boilerplate (#3900)
