+++ packages/admin-panel-server/src/index.ts
@@ -8,9 +8,11 @@ import '@babel/polyfill';
 import * as dotenv from 'dotenv';
 import http from 'http';
 
+import winston from 'winston';
+import { configureWinston } from '@tupaia/server-boilerplate';
 import { createApp } from './app';
-import winston from './log';
 
+configureWinston();
 dotenv.config(); // Load the environment variables into process.env
 
 /**
