+++ packages/server-boilerplate/package.json
@@ -37,6 +37,7 @@
     "i18n": "^0.13.3",
     "express": "^4.16.2",
     "http-proxy-middleware": "^2.0.1",
-    "morgan": "^1.9.0"
+    "morgan": "^1.9.0",
+    "winston": "^3.2.1"
   }
 }
