+++ packages/server-boilerplate/src/utils/index.ts
@@ -3,6 +3,7 @@
  * Copyright (c) 2017 - 2020 Beyond Essential Systems Pty Ltd
  *
  */
+export { configureWinston } from './configureWinston';
 export { handleError } from './handleError';
 export { handleWith } from './handleWith';
 export { useForwardUnhandledRequests } from './useForwardUnhandledRequests';
